feat: validate if specialty exists before execute query to delete in delete user specialty use case

+++ src/modules/user/domain/useCases/delete-user-specialty-use-case.ts
@@ -23,6 +23,12 @@ export class DeleteSpecialtyUseCase {
       throw new AppError('Missing params');
     }
 
+    const userHasSpecialty = await this.specialtyRepository.findOne({ userEmail, name });
+
+    if (!userHasSpecialty) {
+      throw new AppError('Specialty not found');
+    }
+
     const deletedSpecialty = await this.specialtyRepository.delete({ userEmail, name });
 
     return deletedSpecialty;
